Move all code to ee directory
