+++ ee/auth/saml/index.js
@@ -1,8 +1,8 @@
 // @ts-check
 const { MultiSamlStrategy } = require('passport-saml');
 
-const sqldb = require('../../prairielib/lib/sql-db');
-const sqlLoader = require('../../prairielib/lib/sql-loader');
+const sqldb = require('../../../prairielib/lib/sql-db');
+const sqlLoader = require('../../../prairielib/lib/sql-loader');
 
 const sql = sqlLoader.loadSqlEquiv(__filename);
