+++ tsconfig.json
@@ -15,5 +15,6 @@
     "sync/**/*",
     "lib/**/*",
     "tools/**/*",
+    "ee/**/*",
   ],
 }
